requirements.txt
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
idna==3.6
mutagen==1.47.0
pycryptodomex==3.19.0
PyYAML==6.0.1
requests==2.31.0
static-ffmpeg==2.5
urllib3==2.1.0
websockets==12.0
yt-dlp==2023.11.16
